+++ SkillFloors/Package/CHANGELOG.md
@@ -1,11 +1,16 @@
 # Changelog
 
-## 1.1.4
-* fixed a save data migration bug in 1.1.3. Please skip 1.1.3!
+## 1.1.5 (2026-04-21)
+* backend improvements, cleaner execution
+* code cleanup and logging adjustments
+
+## 1.1.4 (2026-04-19)
+* fixed a save data migration bug in 1.1.3. If upgrading from 1.1.2 you should be fine.
 * more verbose debug logging if you turn debug on in SkillFloors config.
 * fixed documentation typos
 
 ## 1.1.3 (2026-04-18)
+* THIS VERSION WILL RESET YOUR SKILL FLOORS IF YOU ARE UPGRADING FROM A PREVIOUS VERSION
 * fixed an issue where new characters could inherit other characters' SkillFloor levels
 * prevents skill floors from being higher their associated skill and won't gain XP if the floor and skill are the same level.
 * moving from JSON to ZPackage saves (deprecated JsonDotNET dependency). Will be removed in 1.2 update.